third_party/WebKit/LayoutTests/external/wpt/cookies/samesite/form-get-blank-expected.txt

This is a testharness.js-based test.
PASS Same-host top-level form GETs are strictly same-site
PASS Subdomain top-level form GETs are strictly same-site
FAIL Cross-site top-level form GETs are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.25687148949670124" but got (undefined) undefined
PASS Same-host redirecting to same-host top-level form GETs are strictly same-site
PASS Subdomain redirecting to same-host top-level form GETs are strictly same-site
PASS Cross-site redirecting to same-host top-level form GETs are strictly same-site
PASS Same-host redirecting to subdomain top-level form GETs are strictly same-site
PASS Subdomain redirecting to subdomain top-level form GETs are strictly same-site
PASS Cross-site redirecting to subdomain top-level form GETs are strictly same-site
FAIL Same-host redirecting to cross-site top-level form GETs are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.7692327300538324" but got (undefined) undefined
FAIL Subdomain redirecting to cross-site top-level form GETs are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.88346224548266" but got (undefined) undefined
FAIL Cross-site redirecting to cross-site top-level form GETs are laxly same-site assert_equals: Non-SameSite cookies are always sent. expected (string) "0.6467834131701868" but got (undefined) undefined
Harness: the test ran to completion.
